Review: Avoid infinite isMetaAnnotated rule exectution
We can shorten the code a little by checking the predicate in the same loop as the meta-annotation. This will effectively switch from a mixture of breath-first-search (check all in the same level for the predicate) and depth-first-search (if the current level fails traverse one branch and repeat) to a pure depth-first-search (only check the predicate on the root of one branch, then immediately descend). But since we have no metrics if one or the other performs better in real life scenarios (and I would also assume "it depends" on the code structure), we can just go for the more concise code.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>